fixing some compile problems on FreeBSD

+++ runtime/cfsysline.c
@@ -217,9 +217,9 @@ static rsRetVal doGetSize(uchar **pp, rsRetVal (*pSetHdlr)(void*, uid_t), void *
 		case 'K': i *= 1000; ++(*pp); break;
 	        case 'M': i *= 1000000; ++(*pp); break;
                 case 'G': i *= 1000000000; ++(*pp); break;
-                case 'T': i *= 1000000000000; ++(*pp); break; /* tera */
-                case 'P': i *= 1000000000000000; ++(*pp); break; /* peta */
-                case 'E': i *= 1000000000000000000; ++(*pp); break; /* exa */
+                case 'T': i *= (long long) 1000000000000; ++(*pp); break; /* tera */
+                case 'P': i *= (long long) 1000000000000000; ++(*pp); break; /* peta */
+                case 'E': i *= (long long) 1000000000000000000; ++(*pp); break; /* exa */
 	}
 
 	/* done */
